requirements.txt
flask
gunicorn
pillow
rembg
opencv-python-headless
numpy